% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtbase.R
\name{crtbase}
\alias{crtbase}
\title{Create base vector}
\usage{
crtbase(Lind, Base=NULL)
}
\arguments{
\item{Lind}{a number or vector containing the lengths
of the alleles. Sum(Lind) is the length of
the corresponding chromosome.}

\item{Base}{a number or vector containing the base of
the loci contained in the Alleles. Default is set to binary
representation.}
}
\value{
a vector whose elements correspond to the base
of the loci of the associated chromosome structure.
}
\description{
This function creates a vector containing the base of the loci
in a chromosome.
}
\examples{

Base = crtbase(10)
}
\author{
The original matlab implementation of mutate was written by Andrew Chipperfield.
The R implementation was written by David Zhao.
}

