% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtbp.R
\name{crtbp}
\alias{crtbp}
\title{Create an initial population}
\usage{
crtbp(Nind, Lind = NULL, Base = NULL)
}
\arguments{
\item{Nind}{- either a number containing the number of individuals
in the new population or a row vector of length two
containing the number of individuals and their length.}

\item{Lind}{an optional number containing the length of the individual
chromosomes.}

\item{Base}{an optional number containing the base of the chromosome 
elements or a row vector containing the base(s) 
of the loci of the chromosomes.}
}
\value{
a list containing following components:
\item{Chrom}{a matrix containing the random valued chromosomes 
row wise.}
\item{Lind}{a scalar containing the length of the chromosome.}
\item{BaseV}{a row vector containing the base of the 
chromosome loci.}
}
\description{
This function creates a binary population of given size and structure.
}
\examples{

Chrom = crtbp(40,10)
}
\author{
The original matlab implementation of mutate was written by Andrew Chipperfield.
The R implementation was written by David Zhao.
}

