% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{mut}
\alias{mut}
\title{MUTATion low-level function}
\usage{
mut(OldChrom,Pm=NULL,BaseV=NULL)
}
\arguments{
\item{OldChrom}{a matrix containing the chromosomes of the
current population. Each row corresponds to
an individuals string representation.}

\item{Pm}{a value indicating mutation probability. Default value
of Pm = 0.7/Lind, where Lind is the chromosome length.}

\item{BaseV}{an optional vector of the same length as the
chromosome structure defining the base of the 
individual elements of the chromosome. Default is set to binary
representation.}
}
\value{
a matrix containing a mutated version of OldChrom.
}
\description{
This function takes the representation of the current population,
mutates each element with given probability and returns the resulting
population.
}
\examples{

Chrom = crtbp(40,10)$Chrom

NewChrom = mut(OldChrom=Chrom)
}
\author{
The original matlab implementation of mutate was written by Andrew Chipperfield.
The R implementation was written by David Zhao.
}

