% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reins.R
\name{reins}
\alias{reins}
\title{RE-INSertion of offspring in population replacing parents}
\usage{
reins(Chrom,SelCh,SUBPOP=1,InsOpt=c(0,1),ObjVCh=NULL,ObjVSel=NULL)
}
\arguments{
\item{Chrom}{matrix containing the individuals (parents) of the current
population. Each row corresponds to one individual.}

\item{SelCh}{matrix containing the offspring of the current
population. Each row corresponds to one individual.}

\item{SUBPOP}{an optional number indicating subpopulations
1 subpopulation is as default.}

\item{InsOpt}{an optional vector containing the insertion method parameters
InsOpt[1] number indicating kind of insertion. 0 - uniform insertion and
1 - fitness-based insertion. if omitted, 0 is assumed.
InsOpt[2] rate of offspring to be inserted per subpopulation (% of subpopulation)
if omitted, 1.0 (100%) is assumed}

\item{ObjVCh}{an optional vector containing the objective values
of the individuals (parents - Chrom) in the current 
population, needed for fitness-based insertion
saves recalculation of objective values for population}

\item{ObjVSel}{an optional vector containing the objective values
of the offspring (SelCh) in the current population, needed for
partial insertion of offspring,
saves recalculation of objective values for population}
}
\value{
a list containing following components:
\item{Chrom}{matrix containing the individuals of the current
 population after reinsertion.}
\item{ObjVCh}{if ObjVCh and ObjVSel are input parameters, then column 
 vector containing the objective values of the individuals
 of the current generation after reinsertion.}
}
\description{
This function reinserts offspring in the population.
}
\author{
The original matlab implementation of bs2rv was written by Hartmut Pohlheim and 
tested by Alex Shenfield. 
The R implementation was written by David Zhao.
}

