% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Universal selection high-level function}
\usage{
select(SEL_F=c("sus","rws"),Chrom,FitnV,GGAP=1,SUBPOP=1,...)
}
\arguments{
\item{SEL_F}{character string indicating the selection function.}

\item{Chrom}{a matrix containing the individuals (parents) of the current
population. Each row corresponds to one individual.}

\item{FitnV}{a vector containing the fitness values of the
individuals in the population.}

\item{GGAP}{an optional number indicating rate of individuals to be selected.
Default is set to 1.}

\item{SUBPOP}{an optional number indicating subpopulations.
Default is set to 1 subpopulation.}

\item{...}{ohter aurguments passed on to selection function.}
}
\value{
a matrix containing the selected individuals.
}
\description{
This function performs universal selection. The function handles
multiple populations and calls the low level selection function
for the actual selection process.
}
\examples{

Chrom = crtbp(40,10)$Chrom

#Ojbective function is the sum of individuals
FitnV = ranking(apply(Chrom,1,sum))

Selch = select(SEL_F="sus",Chrom=Chrom,FitnV=FitnV,GGAP=0.9)
}
\author{
The original matlab implementation of mutate was written by Hartmut Pohlheim.
The R implementation was written by David Zhao.
}

