% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus.R
\name{sus}
\alias{sus}
\title{Stochastic Universal Sampling}
\usage{
sus(FitnV, Nsel)
}
\arguments{
\item{FitnV}{a vector containing the fitness values of the
individuals in the population.}

\item{Nsel}{a number indicating individuals to be selected}
}
\value{
a vector containing the indexes of the selected
individuals relative to the original population, shuffled.
The new population, ready for mating, can be obtained
by calculating OldChrom[NewChrIx,].
}
\description{
This function performs selection with STOCHASTIC UNIVERSAL SAMPLING.
}
\examples{

Chrom = crtbp(40,10)$Chrom

#Ojbective function is the sum of individuals
FitnV = ranking(apply(Chrom,1,sum))

Selch = sus(FitnV=FitnV,Nsel=2)
}
\author{
The original matlab implementation of mutate was written by Hartmut Pohlheim and
Carlos Fonseca.
The R implementation was written by David Zhao.
}

