% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xovmp.R
\name{xovmp}
\alias{xovmp}
\title{Crossover operators for binary representation}
\usage{
xovmp(OldChrom,Px=0.7,Npt=0,Rs=TRUE)
xovsp(OldChrom,Px=0.7)
}
\arguments{
\item{OldChrom}{a matrix containing the chromosomes of the old
population. Each row corresponds to one individual.}

\item{Px}{a number indicating the probability of recombination 
ocurring between pairs of individuals. If omitted, 0.7 is set.}

\item{Npt}{an optinal number indicating the number of crossover points to use.
Specifically, 1 indicates single point recombination; 2 indicates 
double point recombination; and 0 indicates shuffle point recombination.
Default is set to 0.}

\item{Rs}{an optional logical value indicating whether or not to 
force the production of offspring different from their parents.
Default is set to TRUE.}
}
\value{
a matrix containing the offspring after mating, 
ready to be mutated and/or evaluated, in the same 
format as OldChrom.
}
\description{
Taking a matrix OldChrom containing the binary
representation of the individuals in the current population,
applying crossover to consecutive pairs of individuals with
defined probability and returns the resulting population.
}
\examples{

Chrom = crtbp(40,10)$Chrom

Selch = xovmp(OldChrom=Chrom)
}
\author{
The original matlab implementation of mutate was written by Carlos Fonseca and
updated by Andrew Chipperfield.
The R implementation was written by David Zhao.
}

