% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xovsp.R
\name{xovsp}
\alias{xovsp}
\title{CROSSOVer Single-Point}
\usage{
xovsp(OldChrom,XOVR=0.7)
}
\arguments{
\item{OldChrom}{a matrix containing the chromosomes of the old
population. Each line corresponds to one individual.
In any form, not necessarily real values.}

\item{XOVR}{a number indicating the probability of recombination 
occurring between pairs of individuals.}
}
\value{
a matrix containing the chromosomes of the population
after mating, ready to be mutated and/or evaluated,
in the same format as OldChrom.
}
\description{
This function performs single-point crossover between pairs of 
individuals and returns the current generation after mating.
}
\examples{

Chrom = crtbp(40,10)$Chrom

Selch = xovsp(OldChrom=Chrom)
}
\author{
The original matlab implementation of mutate was written by Hartmut Pohlheim.
The R implementation was written by David Zhao.
}

